rename scedc_s3store to s3store
